Created permutations.rs and moved permutation functions into it
